modified: conflict 방지 코드 수정
- ApiResultStatus
- SecurityConfig
- application.yml
