+++ src/main/java/com/example/dm/exception/ApiResultStatus.java
@@ -31,7 +31,6 @@ public enum ApiResultStatus {
     ALREADY_SIGNED_UP("E4090", "이미 가입된 이메일입니다.", HttpStatus.CONFLICT),
     LOGIN_FAILED("E4014", "로그인 정보가 일치하지 않습니다.", HttpStatus.UNAUTHORIZED),
     USER_NOT_FOUND("E4041", "유저 정보가 존재하지 않습니다.", HttpStatus.NOT_FOUND),
-    WRONG_PASSWORD("E4041", "일치하지 않는 비밀번호입니다.", HttpStatus.FORBIDDEN),
 
     // 게시글 관련 에러
     POST_NOT_FOUND("E4042", "존재하지 않는 게시글입니다.", HttpStatus.NOT_FOUND),
