--- src/main/java/com/example/dm/security/SecurityConfig.java
@@ -1,58 +1,0 @@
-package com.example.dm.security;
-
-import java.util.List;
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.CsrfConfigurer;
-import org.springframework.security.config.annotation.web.configurers.HttpBasicConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-@RequiredArgsConstructor
-@Import(PermitUrlProperties.class)
-public class SecurityConfig {
-    @Value("${api.path.default}")
-    private String API_URL_PREFIX;
-    private final PermitUrlProperties urlProperties;
-    private final CustomEntryPoint customEntryPoint;
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.httpBasic(HttpBasicConfigurer::disable)
-            .cors(Customizer.withDefaults())
-            .csrf(CsrfConfigurer::disable)
-            .sessionManagement(config -> config.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-            .authorizeHttpRequests(auth -> {
-                for (String url : urlProperties.getGet()) {
-                    auth.requestMatchers("GET", url).permitAll();
-                }
-                for (String url : urlProperties.getPost()) {
-                    auth.requestMatchers("POST", url).permitAll();
-                }
-                auth.anyRequest().authenticated();
-            })
-            .exceptionHandling(exception -> exception
-                .authenticationEntryPoint(customEntryPoint)
-                .accessDeniedHandler(customEntryPoint)
-            );
-
-        return http.build();
-    }
-}
-
