+++ src/main/resources/application.yml
@@ -12,9 +12,9 @@ spring:
 api:
   path:
     default: /api/v1
-#  jwt:
-#    secret: ${JWT_SECRET}
-#    expire: ${JWT_EXPIRE:600} # seconds
+  jwt:
+    secret: ${JWT_SECRET}
+    expire: ${JWT_EXPIRE:600} # seconds
 
 # aws s3 file server
 aws-s3:
